Add summarize_location and summarize_date (#16)
Variables have date and location information that has varying levels of precision, and can represent multiple sources if they're composite variables. Add methods to summarize these.

```
$ ./summarize.py
hiv
  157.21 per 100k (United States; 2019)
  230.48 per 100k (Los Angeles, California, United States; 2020)
hsv_1
  49600.00 per 100k (United States; 2015)
  47800.00 per 100k (United States; 2015)
  98000.00 per 100k (Louisiana, United States; 2005)
norovirus
  0.82 per 100k (United States; 2012 to 2022)
This is still not perfect: it doesn't tell you that norovirus is trying to be an estimate of the Rothman 2021 study period. Right now that isn't in the code; we need to figure out how to represent it.
```

Also add some missing dates.