prevalence: add a concept of a target (#19)
A prevalence estimate can target a particular location or date, which is not quite the same as being derived from information about particular locations or dates.  Make it explicit when we do have a target.

This PR just adds one for covid; if we like this I can go back and add ones for other pathogens.