norovirus: more detailed prevalence estimate (#53)
Previously I did some manual calculations to see how the Rothman period compared to normal, but this throws away a lot of data that could be useful in modeling. Now I generate one estimates for every month.

Before we were estimating 0.82 per 100k for the Rothman data. With this PR we estimate:

National:
```
  2020-08 | 0.33
  2020-09 | 0.51
  2020-10 | 0.66
  2020-11 | 0.85
  2020-12 | 0.98
  2021-01 | 4.92
```