Hepatitis A: use the taxid the MGS Pipeline assigns reads to
The code had 208726, which was more specific, but Virus-Host DB doesn't know about 208726, only its parent 12092.  Since we don't expect Simian or Marsupial HepA in the sequencing data I think this isn't a problem.

Prompted by trying to make the test more general for #78 and realizing it was failing on some of our pathogens.