summarize: don't skip location and date information from the top-level variable
Before this change, if you made a Variable with a location and then included additional location information on its inputs, only the latter would be considered by summarize_location (and the same with summarize_date).

This was originally included in #85 where it fixed a bug that Flu was triggering, but that PR is taking a long time for review, so pulling it out for separate review (which I probably should have done from the start).