incidence to prevalence: consider incidence over time (#98)
The number of currently infected people today is not today's incidence times the mean shedding duration. It's much closer to the sum of incidences over the last shedding duration days. Do that instead.

Here's how this changes our predictions for LA County. Relatively subtle, but you can see it pushes peaks a bit into the future, and spreads them a bit horizontally, as you'd expect:

Fixes #87