Add stubs for BKV, JCV, MCV, and AAV2 (#162)
We are planning to add estimates for these four viruses (https://twist.com/a/197793/ch/661244/t/4455246/) and it would be helpful to test them in modeling at this stage.  We expect their population prevalence to be constant over time, which means modeling will only be off by a (very important eventually but not important right now) constant factor.